+++ src/prisma/prisma.module.ts
@@ -1,7 +1,7 @@
 // src/prisma/prisma.module.ts
 
 import { Module } from '@nestjs/common';
-import { PrismaService } from './prisma.service';
+import { PrismaService } from './service/prisma.service';
 
 @Module({
   providers: [PrismaService],
