+++ src/prisma/service/prisma.service.ts
@@ -1,5 +1,5 @@
 // src/prisma/prisma.service.ts
-import { INestApplication, Injectable } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { PrismaClient } from '@prisma/client';
 @Injectable()
 export class PrismaService extends PrismaClient {}
